Navigation stack manipulation changed in order to work uniformly across WP8 and Android (iOS remains to be tested).
Software and physical back button invocation is now handled in the same way, where a special back page is inserted in front of the last page in order to be able to detect the software back button invocation. Perhaps the physical back button invocation, through OnBackButtonPressed, will be allowed a shortcut later on, right now it is good to be able to have both back navigation ways treated the same way in order to easily test the functionality on WP8 as well.